requirements.txt
torch>=2.0.0
numpy>=1.20.0
gymnasium>=1.0.0
matplotlib>=3.5.0
tensorboard>=2.10.0
pyyaml>=6.0
tqdm>=4.64.0
scipy>=1.9.0
torch-geometric>=2.3.0
seaborn>=0.11.0
pandas>=1.5.0
